requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
langchain-chroma>=0.1.0
sentence-transformers>=2.2.0
unstructured>=0.14.0
unstructured[pdf]>=0.14.0
langchain-unstructured>=0.0.9
chromadb>=0.4.0
